requirements.txt
matplotlib==3.10.3
numpy==2.2.5
pandas==2.2.3
Pillow==11.2.1
scikit_learn==1.6.1
seaborn==0.13.2
streamlit==1.42.0
torch==2.7.0
torchvision==0.22.0
